flac: fix golangci lint issues
	encode_frame.go:89:1: cyclomatic complexity 52 of func `(*Encoder).encodeFrameHeader` is high (> 30) (gocyclo)
	func (enc *Encoder) encodeFrameHeader(w io.Writer, hdr frame.Header) error {
	^
	internal/utf8/encode.go:66:17: unnecessary conversion (unconvert)
			bits := uint64(tx | (x>>uint(6*i))&maskx)
							  ^
	encode_subframe.go:105:46: unnecessary conversion (unconvert)
			if err := bw.WriteBits(uint64(sample), byte(hdr.BitsPerSample)); err != nil {
																	 ^

+++ internal/utf8/encode.go
@@ -63,7 +63,7 @@ func Encode(w io.Writer, x uint64) error {
 
 	// Store continuation bytes.
 	for i := l - 1; i >= 0; i-- {
-		bits := uint64(tx | (x>>uint(6*i))&maskx)
+		bits := tx | (x>>uint(6*i))&maskx
 		if err := ioutilx.WriteByte(w, byte(bits)); err != nil {
 			return errutil.Err(err)
 		}
